削除ボタン押下時に確認メッセージを表示

+++ admin/includes/view_all_posts.php
@@ -97,7 +97,7 @@ if (isset($_POST['checkBoxArray'])) {
                 echo "<td>$post_comment_count</td>";
                 echo "<td><a href='/cms/post.php?p_id={$post_id}'>View Post</a></td>";
                 echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                echo "<td><a onClick=\"javascript: return confirm('削除しても良いですか？')\" href='posts.php?delete={$post_id}'>Delete</a></td>";
                 echo '</tr>';
             }
             ?>
